Add additional state to Asset component
Assets can be updated via check in and check out. Current
implementation uses asset obtained via route, therefore re-render
is not something easily done when information is added. In the case
of checking out an asset to a user via CheckOut component, assigned
user is updated, and when checking in asset via CheckIn component,
status of asset might be updated.